+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2024-08-06
+
+### Changed
+
+- changed default log-level for reset() and upload() to INFO
+- added a few examples to README.md
+
 ## [0.2.0] - 2024-08-06
 
 ### Added
@@ -35,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 _First release._
 
 
+[0.2.1]: https://github.com/int-brain-lab/tycmd-wrapper/releases/tag/v0.2.1
 [0.2.0]: https://github.com/int-brain-lab/tycmd-wrapper/releases/tag/v0.2.0
 [0.1.2]: https://github.com/int-brain-lab/tycmd-wrapper/releases/tag/v0.1.2
 [0.1.1]: https://github.com/int-brain-lab/tycmd-wrapper/releases/tag/v0.1.1